Use conditional compilation in AudioStream.h to remove AudioData.h
usb_desc.h only needs the sample rates, so we set a temporary macro prior to a nested #include of AudioStream.h, causing the latter to #ifdef out everything but the sample rate information. We also have to be careful *not* to set the AudioStream_h macro when this happens, as we *don't* have the full set of declarations! Seems to work OK...